requirements.txt
docopt==0.6.2
numpy==1.26.4
paho-mqtt==2.1.0
passlib==1.7.4
prettytable==3.16.0
PyYAML==6.0.2
six==1.17.0
tflite-runtime==2.14.0
transitions==0.8.11
wcwidth==0.2.13
websockets==10.4
